Added changes for backend

+++ src/main/java/in/mohit/controller/SftpController.java
@@ -4,6 +4,7 @@ import in.mohit.model.ApiResponse;
 import in.mohit.model.FileModal;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.websocket.server.PathParam;
+import org.apache.catalina.util.StringUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -16,6 +17,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping("/files")
@@ -37,13 +39,12 @@ public class SftpController {
 	}
 
 	@PostMapping(value = "/upload",consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-	public ResponseEntity<String> uploadFile(MultipartFile file, @RequestParam(value = "directory"
+	public ResponseEntity<FileModal> uploadFile(MultipartFile file, @RequestParam(value = "directory"
 			,defaultValue = "/",required = false) String directory) {
 
 		// Get input stream from MultipartFile and upload it to the specified SFTP directory
         try {
-			sftpService.uploadFile(directory, file.getOriginalFilename(), file.getInputStream());
-			return ResponseEntity.status(HttpStatus.CREATED).body("Uploaded the file to SFTP Server");
+			return ResponseEntity.status(HttpStatus.CREATED).body(sftpService.uploadFile(directory, file.getOriginalFilename(), file.getInputStream(),file.getSize()));
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
@@ -53,7 +54,8 @@ public class SftpController {
 	public ResponseEntity<ApiResponse> deleteFile(@RequestParam("filename") String filename
 			, @RequestParam(value = "remoteDirectory",required = false,defaultValue = "/") String remoteDirectory) {
 		try {
-			sftpService.deleteFile(remoteDirectory, filename);
+			if(Objects.nonNull(filename))
+				sftpService.deleteFile(remoteDirectory, filename);
 			return ResponseEntity.ok(new ApiResponse(HttpStatus.OK,HttpStatus.OK.getReasonPhrase(), "Deleted Successfully"));
 		} catch (Exception e) {
 			return ResponseEntity.ok(new ApiResponse(HttpStatus.INTERNAL_SERVER_ERROR,HttpStatus.INTERNAL_SERVER_ERROR.getReasonPhrase(), "File deletion failed"));
